requirements.txt
flask
requests
google-cloud-storage
google-cloud-speech
tenacity
pytest
